+++ etl/src/test/scala/org/astraea/etl/CSVReaderTest.scala
@@ -18,7 +18,10 @@ package org.astraea.etl
 
 import com.opencsv.CSVWriter
 import org.apache.spark.SparkException
-import org.apache.spark.sql.Row
+import org.apache.spark.sql.{Encoder, Row}
+import org.apache.spark.sql.functions.col
+import org.apache.spark.sql.types.StructType
+import org.astraea.etl.CSVReader.{createSpark, csvToJSON}
 import org.astraea.etl.DataType.{IntegerType, StringType}
 import org.junit.jupiter.api.Assertions.assertThrows
 import org.junit.jupiter.api.Test
@@ -154,5 +157,99 @@ class CSVReaderTest {
       .listFiles()
       .filter(!_.isDirectory)
       .filter(t => t.toString.endsWith(".csv"))
+  }
+
+  @Test def csvToJSONTest(): Unit = {
+    val spark = createSpark("local[2]")
+
+    import spark.implicits._
+    case class Person(name: String, age: Long)
+    val data =
+      Seq(Person("Michael", 29), Person("Andy", 30), Person("Justin", 19))
+    implicit val make: Encoder[Person] =
+      org.apache.spark.sql.Encoders.kryo[Person]
+    val df =
+      spark.createDataset(data).map(x => (x.name, x.age)).toDF("name", "age")
+
+    val json = csvToJSON(df, Seq("name"))
+    val iterator = (0 to 2).iterator
+    json
+      .collectAsList()
+      .forEach(row => {
+        val i = iterator.next()
+        assert(row(0) equals data(i).name)
+        assert(
+          row(1) equals s"""{"name":"${data(i).name}","age":${data(i).age}}"""
+        )
+      })
+  }
+
+  @Test def csvToJsonMulKeysTest(): Unit = {
+    val spark = createSpark("local[2]")
+
+    import spark.implicits._
+    case class Person(firstName: String, secondName: String, age: Long)
+    val data =
+      Seq(
+        Person("Michael", "A", 29),
+        Person("Andy", "B", 30),
+        Person("Justin", "C", 19)
+      )
+    implicit val make: Encoder[Person] =
+      org.apache.spark.sql.Encoders.kryo[Person]
+    val df =
+      spark
+        .createDataset(data)
+        .map(x => (x.firstName, x.secondName, x.age))
+        .toDF("firstName", "secondName", "age")
+
+    val json = csvToJSON(
+      df,
+      Seq("firstName", "secondName")
+    )
+    val iterator = (0 to 2).iterator
+    json.show()
+    json
+      .collectAsList()
+      .forEach(row => {
+        val i = iterator.next()
+        assert(row(0) equals s"${data(i).firstName}${data(i).secondName}")
+        assert(
+          row(1) equals
+            s"""{"firstName":"${data(
+              i
+            ).firstName}","secondName":"${data(i).secondName}","age":${data(
+              i
+            ).age}}"""
+        )
+      })
+  }
+
+  @Test def jsonToByteTest(): Unit = {
+    val spark = createSpark("local[2]")
+
+    var data = Seq(Row(1, "A1", 52, "fghgh", "sfjojs", "zzz", "final", 5))
+    (0 to 10000).iterator.foreach(_ =>
+      data = data ++ Seq(Row(1, "A1", 52, "fghgh", "sfjojs", "zzz", "final", 5))
+    )
+
+    val structType = new StructType()
+      .add("ID", "integer")
+      .add("name", "string")
+      .add("age", "integer")
+      .add("xx", "string")
+      .add("yy", "string")
+      .add("zz", "string")
+      .add("f", "string")
+      .add("fInt", "integer")
+
+    val df =
+      spark.createDataFrame(spark.sparkContext.parallelize(data), structType)
+
+    val json = csvToJSON(df, Seq("ID"))
+      .withColumn("byte", col("value").cast("Byte"))
+      .selectExpr("CAST(byte AS BYTE)")
+    val head = json.head()
+    assert(json.filter(_ != head).isEmpty)
   }
 }
